btree: improve leaf cell payload handling
Improves the in-memory representation of the leafcell struct. Now,
writing to disk just involves using the std::arrays and cellSize() to
find the actual cell size.

Although using a fixed size array for smallPayload, when this could be
smaller than MAX_PAYLOAD_SIZE, is strange, it is easier to use than a
complex struct; this way, we get to keep default constructors. Just make
sure to use cellSize() when writing.