add interior/leaf nodes; leaf searching for value
searchGetLeaf will return the leaf node a value would be in. It may or
may not exist there, and will not point to where in that node it would
be.

This will be used for all the other types of searching, as it gives
the starting point in the tree of where to search. To validate the
existence of the search value, you do not need to traverse the leaf
linked list, as it will have found the correct leaf node.