slots: use templated buffer size and `this+1`
When the page is set into the pager, the span pointer is no longer
valid.

This could be fixed in other ways, but we were still storing the span
object in the struct which wasn't ideal.

The SlotHeader now cannot be created by passing in the span, it must
live inside the buffer itself and be created using the available space
in the buffer as a template parameter.

The only place SlotHeader is used is in the BTreeHeader, but this could
change. I probably should have kept the code in the header file.